[plugin.video.newsy] 1.0.0
