+++ plugin.video.newsy/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<addon id="plugin.video.newsy" version="1.0.0" name="Newsy" provider-name="Lunatixz">
+<addon id="plugin.video.newsy" version="1.0.1" name="Newsy" provider-name="Lunatixz">
     <requires>
         <import addon="xbmc.python" version="2.25.0"/>
         <import addon="script.module.simplecache" version="1.0.0"/>
@@ -17,7 +17,7 @@
         <source>https://github.com/Lunatixz/XBMC_Addons/tree/master/plugin.video.newsy</source>
         <website>http://www.newsy.com</website>
         <forum>https://forum.kodi.tv/showthread.php?tid=320402</forum>
-        <news></news>
+        <news>[CR]v1.0.1[CR]- Improved browsing speed.</news>
         <assets>
             <icon>resources/images/icon.png</icon>
             <fanart>resources/images/fanart.jpg</fanart>
